+++ vite.config.ts
@@ -13,5 +13,11 @@ export default defineConfig({
   build: {
     target: 'es2022',
     sourcemap: true,
+    rollupOptions: {
+      input: {
+        main: 'index.html',
+        performance: 'performance.html',
+      },
+    },
   },
 });